refactor(errors): move code registry below storage layer
