+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "crates/atm-architecture",
+    "crates/atm-error",
     "crates/atm-storage",
     "crates/atm-storage-sqlserver-proof",
     "crates/atm-core",
