+++ crates/atm-core/Cargo.toml
@@ -20,6 +20,7 @@ name = "atm_core"
 test-utils = []
 
 [dependencies]
+atm-error = { path = "../atm-error", version = "1.3.2-beta.25" }
 atm-storage = { path = "../atm-storage", version = "1.3.2-beta.25" }
 serde.workspace = true
 serde_json.workspace = true
