+++ crates/atm-error/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "atm-storage"
+name = "atm-error"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -7,12 +7,10 @@ authors.workspace = true
 license.workspace = true
 repository.workspace = true
 homepage.workspace = true
-description = "Shared audited storage contract and canonical domain types for ATM."
+description = "Shared ATM error-code vocabulary used by storage and service layers."
 
 [dependencies]
-atm-error = { path = "../atm-error", version = "1.3.2-beta.25" }
 serde.workspace = true
+
+[dev-dependencies]
 serde_json.workspace = true
-chrono = { version = "0.4", features = ["serde"] }
-toml = "0.8"
-ulid.workspace = true
